requirements.txt
click==8.1.5
cmake==3.26.4
filelock==3.12.2
Jinja2==3.1.2
joblib==1.3.1
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pandas==2.0.3
pyarrow==12.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
